+++ imagepicker/src/main/java/com/akshayashokcode/imagepicker/util/FileUtils.kt
@@ -5,14 +5,14 @@ import android.net.Uri
 import androidx.core.content.FileProvider
 import java.io.File
 
-object FileUtils {
+internal object FileUtils {
     fun createTempImageUri(context: Context): Uri? {
         return try {
             val file = File.createTempFile("camera_image_", ".jpg", context.cacheDir).apply {
                 createNewFile()
                 deleteOnExit()
             }
-            FileProvider.getUriForFile(context, "${context.packageName}.fileprovider", file)
+            FileProvider.getUriForFile(context, "${context.packageName}.imagepicker.provider", file)
         } catch (e: Exception) {
             e.printStackTrace()
             null